Add a first test
